[fix] Creating DeviceReader class and methods moving
